Fixed mark

+++ server/src/main/java/com/ftf/volunteeringproject/volunteer/Volunteer.java
@@ -33,7 +33,7 @@ public class Volunteer {
 
     @Column(nullable = false)
     private String password;
-    private Long mark;
+    private int mark;
 
     private boolean enabled = false;
 }
